Add MAFs and arbitrary CHASM columns to excel
Add a python script that combines strelka_varscan indels and mutect snv into a
single excel workbook. New features are the additon of TUMOR/NORMAL MAFs and
arbitrary CHASM score columns to the summary. The script can also merge the
mutation tsvs with the absolute cancer cell fractions resulting in an extra
column cancer cell fraction for all sheets in the summary output.